Revert "Add key warning to nested collections"
This heuristic isn't great because it relies on inspecting deep children which aren't guaranteed to be React elements. In particular, this was causing stack overflows in a component we had that used a *DOM node* as children, like `<DOMContainer>{node}</DOMContainer>`.

This reverts commits:
3622c25580a2d40c9d63fc51a688736c332e1b9d
d60e804f614877b3d17be905fb71f5c6769ff20e
1d62b0cdc295e4664e202a2f6806de6c1343d36e
1988bf75864db939cc4999fba61b276d84cae901
754d521ef146e02ce9a98a38739a1388dd59c61d