Check warning/invariant calls have good messages
The `warning` implementation checked that the message is long enough to
be useful. See commit 01806e573 for more information.

It makes more sense to move this into a lint rule, and also to apply it
for both `warning` and `invariant`. We can safely remove stuff from the
`warning` implementation as we replace the function internally anyways.

https://github.com/facebook/react/pull/4021#issuecomment-108694976
@spicyj